modules/agentes/Archon/pyproject.toml
[project]
name = "Pandora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
]
